Move title+description inside the paper
